Minor formatting, cleanup
