+++ package.json
@@ -88,7 +88,7 @@
   "dependencies": {
     "@fortawesome/fontawesome-free": "^5.14.0",
     "@hms-dbmi-bgm/react-workflow-viz": "0.1.4",
-    "@hms-dbmi-bgm/shared-portal-components": "github:4dn-dcic/shared-portal-components#0.1.27",
+    "@hms-dbmi-bgm/shared-portal-components": "github:4dn-dcic/shared-portal-components#0.1.28",
     "auth0-lock": "^11.30.4",
     "d3": "^5.16.0",
     "detect-browser": "^3.0.1",
