+++ src/encoded/static/components/item-pages/CaseView/FilteringTableFilterSetUI/index.js
@@ -8,8 +8,8 @@ import memoize from 'memoize-one';
 import ReactTooltip from 'react-tooltip';
 
 import { console } from '@hms-dbmi-bgm/shared-portal-components/es/components/util';
-import { AboveTableControlsBase } from '@hms-dbmi-bgm/shared-portal-components/es/components/browse/components/above-table-controls/AboveTableControlsBase';
 
+import { AboveTableControlsBaseCGAP } from './../../../browse/AboveTableControlsBaseCGAP';
 import { AddToVariantSampleListButton } from './AddToVariantSampleListButton';
 import { SaveFilterSetButton } from './SaveFilterSetButton';
 import { SaveFilterSetPresetButton } from './SaveFilterSetPresetButton';
@@ -187,6 +187,7 @@ export class FilteringTableFilterSetUI extends React.PureComponent {
             // From EmbeddedSearchView:
             context: searchContext, // Current Search Response (not that of this filterSet, necessarily)
             hiddenColumns, addHiddenColumn, removeHiddenColumn, columnDefinitions,
+            sortBy, sortColumns, // From SPC/SortController
             requestedCompoundFilterSet, // From SPC/VirtualHrefController
 
             // From FilteringTab (& higher, e.g. App/redux-store):
@@ -278,8 +279,7 @@ export class FilteringTableFilterSetUI extends React.PureComponent {
 
         if (isActiveDotRouterTab) {
             aboveTableControls = (
-                <AboveTableControlsBase {...{ hiddenColumns, addHiddenColumn, removeHiddenColumn, columnDefinitions }}
-                    panelMap={AboveTableControlsBase.getCustomColumnSelectorPanelMapDefinition(this.props)}>
+                <AboveTableControlsBaseCGAP {...{ hiddenColumns, addHiddenColumn, removeHiddenColumn, columnDefinitions, sortBy, sortColumns }}>
                     <h4 className="text-400 col-12 col-lg my-0 py-1">
                         <strong className="mr-1">{ totalCount }</strong>
                         <span>
@@ -303,7 +303,7 @@ export class FilteringTableFilterSetUI extends React.PureComponent {
                             : null }
                         { (searchType === "VariantSample") && <ExportSearchSpreadsheetButton {...{ requestedCompoundFilterSet, caseItem }} /> }
                     </div>
-                </AboveTableControlsBase>
+                </AboveTableControlsBaseCGAP>
             );
         }
