+++ src/encoded/static/components/navigation/components/CollapsedNav.js
@@ -71,19 +71,21 @@ function HelpNavItem(props){
  * @todo Test user actions or role for things to have here?
  */
 function LeftNavAuthenticated(props){
-    const { context, href } = props;
-    const isCasesLinkActive = useMemo(function(){
-        const { "@id": contextID } = context;
-        const { query = {} } = url.parse(href || contextID, true);
-        return query.type === 'Case';
-        // We assume href and context change together, so we memoize on context instead of href
-        // since is a more performant comparison.
-    }, [ context ]);
+    //const { context, href } = props;
+    // const isCasesLinkActive = useMemo(function(){
+    //     const { "@id": contextID } = context;
+    //     const { query = {} } = url.parse(href || contextID, true);
+    //     return query.type === 'Case';
+    //     // We assume href and context change together, so we memoize on context instead of href
+    //     // since is a more performant comparison.
+    // }, [ context ]);
     return (
         <div className="navbar-nav mr-auto">
+            {/*
             <a href="/search/?type=Case&proband_case=true" className={"nav-link browse-nav-btn" + (isCasesLinkActive ? " active" : "")}>
                 Cases
             </a>
+            */}
             <HelpNavItem {...props} />
         </div>
     );
