+++ src/encoded/static/components/util/PatchItemsProgress.js
@@ -1,6 +1,6 @@
 'use strict';
 
-import React, { useCallback, useEffect, useMemo, useState } from 'react';
+import React from 'react';
 import ReactTooltip from 'react-tooltip';
 import Modal from 'react-bootstrap/esm/Modal';
 import { ajax, console } from '@hms-dbmi-bgm/shared-portal-components/es/components/util';
