+++ src/encoded/static/scss/encoded/_bootstrap-theme-overrides.scss
@@ -211,7 +211,7 @@ $margin-util-sizes-mini: 1, 2, 3, 4, 5, 6, 7, 8; 		// * 1px
     border-radius: 3px;
 
     // For newer browsers, ensure doesn't go beyond window height
-    max-height: 75vh;
+    max-height: 50vh;
     overflow-y: auto;
 }
