+++ src/encoded/static/scss/encoded/_layout.scss
@@ -160,19 +160,17 @@ body {
 
 @mixin make-wide-container {
 	width: auto;
-	padding-left: 12px;
-	padding-right: 12px;
+	padding-left: 1rem; /* == 16px */
+	padding-right: 1rem;
 	@include media-breakpoint-between(md, lg) {
-		width: auto;
 		max-width: none;
-		padding-left: 24px;
-		padding-right: 24px;
+		padding-left: 1.5rem;
+		padding-right: 1.5rem;
 	}
 	@include media-breakpoint-up(xl) {
-		width: auto;
 		max-width: none;
-		padding-left: 36px;
-		padding-right: 36px;
+		padding-left: 2rem;
+		padding-right: 2rem;
 	}
 }
